refactor(matrix-builder, asset-velocity): extract helpers + types into subfolders
